Refactor the share items
